Move utils code
